Beach defender 2.1
This version of my code has trash that increases its falling speed every time you score 10 points. In this version you can also play the game again as the score, you're lives reset. This still isn't fully complete as the speed the trash falls does not reset so when you go to play the game again the trash is still falling very fast